requirements.txt
# requirements.txt
anyio==4.3.0
appdirs==1.4.4
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
h11==0.14.0
httpcore==1.0.4
httpx==0.25.2
idna==3.6
python-telegram-bot==20.6
requests==2.31.0
sniffio==1.3.1
soupsieve==2.5
urllib3==2.2.1
